pyproject.toml
[tool.poetry]
name = "tinfoil"
version = "0.1.0"
description = ""
authors = ["jag-k"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
aiofile = "^3.8.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
black = {extras = ["d"], version = "^23.3.0"}
ruff = "^0.0.260"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | \.venv
)/
'''


[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["*locales*"]
select = [
    "E", # pyflakes
    "F", # pycodestyle errors
    "W", # pycodestyle warnings
    "UP", # pyupgrade
    "I", # isort
    "C4", # flake8-comprehensions
    # pytest
    "PT018", # Assertion should be broken down into multiple parts
    "PT022", # No teardown in fixture {name}, use return instead of yield
]
ignore = ["E501"]
